+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Version 1.4.x
 - Setting `--engine dd-to-sparse --bisimulation` now triggers extracting the sparse bisimulation quotiont
 - JIT model building is now invoked via `--engine jit` (instead of `--jit`)
 - DRN: support import of choice labelling
+- Added option `--build:buildchoiceorig` to build a model (PRISM or JANI) with choice origins (which are exported with, e.g. `--exportscheduler`).
+- Apply the maximum progress assumption while building a Markov automata with the Dd engine.
+- Added option `--build:nomaxprog` to disable applying the maximum progress assumption during model building (for Markov Automata)
 - `storm-pomdp`: Only accept POMDPs that are canonical
 - `storm-pomdp`: Prism language extended with observable expressions
 - `storm-pomdp`: Various fixes that prevented usage.